* Rename gaia to gaiad pack in the app
