+++ infra/apps/gaiad/gaia.go
@@ -1,4 +1,4 @@
-package gaia
+package gaiad
 
 import (
 	"bytes"
