+++ infra/apps/gaiad/wallets.go
@@ -1,4 +1,4 @@
-package gaia
+package gaiad
 
 // RelayerMnemonic is mnemonic used be the relayer.
 const RelayerMnemonic = "gas december mango eager element proof budget polar layer worth there eight delay conduct ring wing hover fury flip shield task dismiss ahead olive"
